requirements.txt
backports.tarfile==1.2.0
dj-database-url==3.1.0
django-cors-headers==4.9.0
django-extensions==4.1
django-jazzmin==3.0.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
openai==2.15.0
packaging==24.2
pillow==12.1.0
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2-binary==2.9.11
python-decouple==3.8
tomli==2.0.1
whitenoise==6.11.0
gunicorn
django-crontab
pywebpush
cryptography
